refactor: outsource parts of core to extra files
