+++ src/settingsModal.tsx
@@ -1,26 +1,22 @@
 import awaitImports from '#lib/imports';
-import { getLoadingSpinner } from '#lib/DOM';
+import { ChangelogBtn } from '#core/changelog';
 import globalStyle from '#style/index.module.scss';
 import { Modal } from '#lib/Modal';
 import { require } from '#lib/require.js';
-import { lt as semverLt } from '#lib/semver';
 import settingsStyle from '#style/settings.module.scss';
 import { STORAGE_V2_SEEN_SETTINGS_KEY } from './migrateStorage';
 import TempStorage from '#lib/TempStorage';
 import type { ThemeBoostBootstrapTooltipClass } from '#types/require.js/theme_boost/bootstrap/tooltip.d.ts';
 import toast from '#lib/toast';
 import { BETTER_MOODLE_LANG, LL } from '#i18n';
-import { cachedRequest, NETWORK_CACHE_KEY, request } from '#lib/network';
+import { cachedRequest, NETWORK_CACHE_KEY } from '#lib/network';
+import { checkForUpdates, VersionBox } from '#core/updateCheck';
 import {
     debounce,
-    htmlToElements,
     isNewInstallation,
-    isNightly,
-    mdID,
     mdToHtml,
     rawGithubPath,
 } from '#lib/helpers';
-import { FIVE_MINUTES, ONE_DAY, ONE_MINUTE, ONE_SECOND } from '#lib/times';
 import {
     GithubLink,
     NavbarItem,
@@ -30,7 +26,8 @@ import {
     highlightNewSettings as highlightNewSettingsSetting,
     newSettingsTooltip as newSettingsTooltipSetting,
     updateNotification as updateNotificationSetting,
-} from './features/general';
+} from '#feats/general';
+import { ONE_DAY, ONE_SECOND } from '#lib/times';
 
 const seenSettings = GM_getValue<string[]>(STORAGE_V2_SEEN_SETTINGS_KEY, []);
 
@@ -53,64 +50,9 @@ const SettingsBtn = (
         </div>
     </NavbarItem>
 ) as NavbarItemComponent;
-const UpdateAvailableBadge = (
-    <div className="count-container"></div>
-) as HTMLDivElement;
 
 // append the Button to the navbar
 SettingsBtn.put();
-// endregion
-
-// region changelog button
-const changelogPath = `/blob/${__GITHUB_BRANCH__}/CHANGELOG.md`;
-const ChangelogBtn = (
-    <GithubLink
-        path={changelogPath}
-        icon={false}
-        className="btn btn-outline-primary"
-        title={LL.settings.changelog()}
-    >
-        <i className="fa fa-history fa-fw"></i>
-        <span>{LL.settings.changelog()}</span>
-    </GithubLink>
-);
-
-const changelogIdPrefix = 'changelog';
-
-/**
- * Fetches the changelog from the GitHub repo and converts it to HTML.
- * Uses the cached HTML if it is not older than 5 minutes.
- * @returns the HTML string of the changelog
- */
-const getChangelogHtml = () =>
-    cachedRequest(rawGithubPath('CHANGELOG.md'), FIVE_MINUTES, 'text', md =>
-        mdToHtml(
-            md
-                // remove the title
-                .replace(/^#\s.*/g, '')
-                // add a horizontal rule before each heading except first
-                .trim()
-                .replace(/(?<=\n)(?=^##\s)/gm, '---\n\n'),
-            3,
-            changelogIdPrefix
-        )
-    ).then(({ value }) => value);
-
-ChangelogBtn.addEventListener('click', e => {
-    e.preventDefault();
-    new Modal({
-        type: 'ALERT',
-        large: true,
-        title: (
-            <>
-                <GithubLink path={changelogPath} /> Better-Moodle:&nbsp;
-                {LL.settings.changelog()}
-            </>
-        ),
-        body: getChangelogHtml(),
-        removeOnClose: true,
-    }).show();
-});
 // endregion
 
 // region support button and information
@@ -152,117 +94,15 @@ SupportBtn.addEventListener('click', e => {
     }).show();
 });
 
-const latestVersionEl = (<code></code>) as HTMLElement;
-
 const SupportWrapper = (
     <div
         id={settingsStyle.supportWrapper}
         className="d-flex flex-row small card border-light mb-3 flex-wrap flex-lg-nowrap"
     >
         {SupportBtn}
-        <div className="d-flex flex-row w-100 align-items-center justify-content-around">
-            <span>
-                {LL.settings.modal.installedVersion()}: <br />
-                {isNightly ? '🌜️ ' : ''}
-                <code>{GM_info.script.version}</code>
-            </span>
-            <span>
-                {LL.settings.modal.latestVersion()}: <br />
-                {isNightly ? '🌜️ ' : ''}
-                {latestVersionEl}
-            </span>
-        </div>
+        {VersionBox}
     </div>
 );
-
-const UpdateBtn = (
-    <button className="btn btn-primary btn-sm col-lg-3">
-        {LL.update.btn()}
-    </button>
-);
-UpdateBtn.addEventListener('click', e => {
-    e.preventDefault();
-
-    new Modal({
-        type: 'SAVE_CANCEL',
-        title: LL.update.title(),
-        body: getChangelogHtml().then(changelogHtml => {
-            const body = <></>;
-            body.append(
-                ...Array.from(htmlToElements(mdToHtml(LL.update.body()))),
-                ...Array.from(htmlToElements(changelogHtml))
-            );
-            const currentId = mdID(
-                `* ${GM_info.script.version}`,
-                changelogIdPrefix
-            );
-            body.querySelectorAll(
-                `[id^="${currentId}"], [id^="${currentId}"] ~ *`
-            ).forEach(el => el.remove());
-            return body;
-        }),
-        buttons: { save: LL.update.reload(), cancel: LL.update.close() },
-        removeOnClose: true,
-    })
-        .onSave(() => window.location.reload())
-        .on('bodyRendered', () => {
-            if (GM_info.script.downloadURL) {
-                open(GM_info.script.downloadURL, '_self');
-            }
-        })
-        .show();
-});
-
-let updateCheckRetryTimeout: ReturnType<(typeof window)['setTimeout']> | null;
-
-/**
- * Checks if there is a newer version of Better-Moodle available.
- * If yes, an update button is added to settings modal.
- * The latest available version is shown in settings modal.
- * If enabled, a red dot is appended to the settings trigger button.
- * @returns void
- */
-const checkForUpdates = () =>
-    getLoadingSpinner('settings')
-        .then(spinner => latestVersionEl.replaceChildren(spinner))
-        .then(() =>
-            request(
-                `https://api.github.com/repos/${__GITHUB_USER__}/${__GITHUB_REPO__}/releases/${isNightly ? 'tags/nightly' : 'latest'}`
-            )
-        )
-        .then(res => res.json())
-        .then(({ tag_name: latestVersion }: { tag_name: string }) => {
-            if (!latestVersion) {
-                throw new Error(
-                    `It is unlikely that ${JSON.stringify(latestVersion)} is the latest version. Aborting update check, please try again in a minute.`
-                );
-            }
-
-            latestVersionEl.replaceChildren(latestVersion);
-
-            return semverLt(GM_info.script.version, latestVersion);
-        })
-        .then(updateAvailable => {
-            updateCheckRetryTimeout = null;
-            if (!updateAvailable) {
-                UpdateAvailableBadge.remove();
-                return;
-            }
-            document
-                .getElementById(settingsStyle.supportWrapper)
-                ?.append(UpdateBtn);
-            if (updateNotificationSetting.value) {
-                document
-                    .getElementById(settingsStyle.openSettingsBtn)
-                    ?.append(UpdateAvailableBadge);
-            } else UpdateAvailableBadge.remove();
-        })
-        .catch(() => {
-            updateCheckRetryTimeout ??= setTimeout(
-                () => void checkForUpdates(),
-                ONE_MINUTE
-            );
-        });
 
 void checkForUpdates();
 updateNotificationSetting.onChange(() => void checkForUpdates());
